requirements.txt
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
plotly==5.17.0
scipy==1.11.4
streamlit==1.28.1
fpdf2==2.7.5
pingouin==0.5.3
openpyxl==3.1.2